Crate zng_wgt_stack

Stack widgets, properties and nodes.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 2 feature flags, 1 enabled by default.

"debug_default"

Enables the "dyn_closure" in debug builds.

Enabled by default.

"dyn_closure"

Box closures at opportune places, such as Var::map, reducing the number of monomorphised types.

This speeds-up compilation time at the cost of runtime.

Structs

Stack

W Stack layout.

StackDirection

Defines a placement point in the previous item and the origin point of the next.

Traits

WidgetInfoStackExt

Extension methods for WidgetInfo that may represent a Stack! instance.

Functions

children

c Stack items.

children_align

c Items alignment.

direction

c Stack direction.

get_index

P Get the child index in the parent stack.

get_index_len

P Get the child index and number of children.

get_rev_index

P Get the child index, starting from the last child at 0.

is_even

P If the child index is even.

is_first

P If the child is the first.

is_last

P If the child is the last.

is_odd

P If the child index is odd.

lazy_sample

Create a node that estimates the size of stack panel children.

lazy_size

Create a node that estimates the size of stack panel children.

node

Stack node.

spacing

c Space in-between items.

stack_nodes

Basic Z-stack node sized by one of the items.