zng_wgt_stack
Function direction

pub fn direction(direction: impl IntoVar<StackDirection>)

c Stack direction.

Capture-Only

This property is capture-only, it only defines a property signature, it does not implement any behavior by itself. Widgets can capture and implement this property as part of their intrinsics, otherwise it will have no effect if set on a widget that does not implement it.