Crate zng_wgt_stack

This crate is part of the zng project.

Stack widgets, properties and nodes.

Structs

Stack

W Stack layout.

StackDirection

Defines a placement point in the previous item and the origin point of the next.

Traits

WidgetInfoStackExt

Extension methods for WidgetInfo that may represent a Stack! instance.

Functions

children

c Stack items.

children_align

c Items alignment.

direction

c Stack direction.

get_index

P Get the child index in the parent stack.

get_index_len

P Get the child index and number of children.

get_rev_index

P Get the child index, starting from the last child at 0.

is_even

P If the child index is even.

is_first

P If the child is the first.

is_last

P If the child is the last.

is_odd

P If the child index is odd.

lazy_sample

Create a node that estimates the size of stack panel children.

lazy_size

Create a node that estimates the size of stack panel children.

node

Stack node.

spacing

c Space in-between items.

stack_nodes

Basic z-stack node.

stack_nodes_layout_by

Basic z-stack node sized by one of the items.